requirements.txt
streamlit==1.29.0
PyMuPDF==1.23.8
Pillow==10.1.0
pytesseract==0.3.10